Split project-wizard.tsx into feature-scoped components
The 1650-line, ~69-useState wizard mixed 7 steps, several conditional
sub-panels, and the environment-creation payload builder in one
component. Split into src/features/project-wizard/, mirroring the
earlier environment-window.tsx decomposition's directory shape.

Unlike environment-window.tsx, the wizard's state isn't one object
(many fields aren't Environment fields at all), so this split keeps
all state flat in the orchestrator and threads same-named values +
setters as props to each extracted step - a mechanical cut-paste-wrap,
not a reducer rewrite. The environment-creation payload literal
becomes a pure buildEnvironmentPayload() function taking an input type
named after the wizard's own local variables, so the call site stays
plain shorthand.

No behavior change intended - same steps, same validation, same
Tauri commands.
