+++ src/features/project-wizard/helpers.ts
@@ -1,0 +1,76 @@
+import type { WizardEnvironmentPayload, WizardPayloadInput } from "./types"
+
+export function buildEnvironmentPayload(input: WizardPayloadInput): WizardEnvironmentPayload {
+  return {
+    id: input.id,
+    name: input.name,
+    webServer: input.webServer,
+    webVersion: input.webServer === "Nginx" ? input.webVersion : "2.4",
+    phpVersion: input.phpVersion,
+    database: input.database,
+    databaseVersion: input.databaseVersion,
+    port: input.nodePort,
+    webContainerName: "",
+    databaseContainerName: "",
+    phpExtensions: input.phpExtensions,
+    extraServices: input.isNative
+      ? input.services.filter((item) => item === "node")
+      : input.services,
+    nodeVersion: input.nodeVersion,
+    redisVersion: "7.4",
+    databaseName: input.databaseName,
+    databaseUser: input.databaseUser,
+    databasePassword: input.databasePassword,
+    databaseRootPassword: input.databaseRootPassword,
+    phpMemoryLimit: input.phpMemoryLimit,
+    phpUploadLimit: input.phpUploadLimit,
+    phpPostLimit: input.phpUploadLimit,
+    phpExecutionTime: input.phpExecutionTime,
+    phpJit: input.phpJit,
+    phpJitMode: input.phpJitMode,
+    phpJitBufferSize: input.phpJitBufferSize,
+    phpCron: input.phpCron,
+    phpCronSchedule: input.phpCronSchedule,
+    phpCronCommand: input.phpCronCommand,
+    phpFpmProcessManager: input.phpFpmProcessManager,
+    phpFpmMaxChildren: input.phpFpmMaxChildren,
+    phpFpmStartServers: input.phpFpmStartServers,
+    phpFpmMinSpareServers: input.phpFpmMinSpareServers,
+    phpFpmMaxSpareServers: input.phpFpmMaxSpareServers,
+    phpFpmMaxRequests: input.phpFpmMaxRequests,
+    phpXdebug: input.phpXdebug,
+    phpXdebugMode: input.phpXdebugMode,
+    phpXdebugPort: input.phpXdebugPort,
+    phpXdebugStart: input.phpXdebugStart,
+    phpXdebugIdeKey: input.phpXdebugIdeKey,
+    environmentVariables: {
+      APP_ENV: input.appEnvironment,
+      DB_CHARSET: input.databaseEncoding,
+      LS_PANEL_AUTO_CREATE_DATABASE: String(input.autoCreateDatabase),
+      ...(input.sqlDump ? { LS_PANEL_SQL_DUMP: input.sqlDump } : {}),
+      NODE_INSTALL_COMMAND: input.nodeInstallCommand,
+    },
+    redisPassword: "",
+    redisMemoryLimit: "256mb",
+    redisEvictionPolicy: "allkeys-lru",
+    nodePackageManager: input.nodePackageManager,
+    nodeAutoInstall: true,
+    nodeAutoRestart: input.nodeAutoRestart,
+    nodeCommand: input.nodeCommand,
+    nodeRunMode: input.nodeRunMode,
+    nodeDevCommand: input.nodeDevCommand,
+    nodeBuildCommand: input.nodeBuildCommand,
+    nodeStartCommand: input.nodeStartCommand,
+    nodeInspector: input.nodeInspector,
+    nodeInspectorPort: input.nodeInspectorPort,
+    runtimeMode: input.executionMode,
+    composerVersion: input.composerVersion,
+    restartPolicy: "unless-stopped",
+    cpuLimit: "2.0",
+    containerMemoryLimit: "2g",
+    wordpressSiteTitle: input.wordpressSiteTitle,
+    wordpressAdminUser: input.wordpressAdminUser,
+    wordpressAdminPassword: input.wordpressAdminPassword,
+    wordpressAdminEmail: input.wordpressAdminEmail,
+  }
+}
